RDoc-2474 Client API > Session > Querying > Text Search - Part 12 (ExactMatch, C#)
